Fixed return type for rateScore

+++ eo-dispatcher/src/main/java/com/dla/foundation/fis/eo/entities/FISEvent.java
@@ -192,7 +192,7 @@ public class FISEvent implements Serializable {
 		this.purchaseTimestamp = purchaseTimestamp;
 	}
 
-	public double getRateScore() {
+	public int getRateScore() {
 		return rateScore;
 	}
